html/browsers/browsing-the-web/history-traversal/ScrollRestorationOrder.html

<!DOCTYPE html>
<style>
  body {
    height: 200vw;
    width: 200vw;
  }
</style>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  'use strict';


  async_test(function(t) {
    window.addEventListener("load", t.step_func(function() {
      // create history entries and then verify the impact of scrollRestoration
      // when they are popped
      assert_array_equals([window.pageXOffset, window.pageYOffset], [0, 0]);

      history.pushState('#1', '', '#1');
      window.scrollTo(50, 100);
      assert_array_equals([window.pageXOffset, window.pageYOffset], [50, 100]);

      history.pushState('#2', '', '#2');
      window.scrollTo(100, 200);
      assert_array_equals([window.pageXOffset, window.pageYOffset], [100, 200]);

      // reset the scroll and kick off the verification
      setTimeout(function() {
        history.pushState(null, null, '#done');
        window.scrollTo(555, 555);
        assert_array_equals([window.pageXOffset, window.pageYOffset], [555, 555]);
        window.history.back();
      }, 0);
    }));

    window.addEventListener('popstate', t.step_func(function() {
      var key = location.hash;

      switch (key) {
        case '#2':
          assert_equals(history.state, key, 'state is not ' + key);
          assert_array_equals([window.pageXOffset, window.pageYOffset], [100, 200], 'scroll is not restored before popstate for ' + key);
          break;
        case '#1':
          assert_equals(history.state, key, 'state is not ' + key);
          assert_array_equals([window.pageXOffset, window.pageYOffset], [50, 100], 'scroll is not restored before popstate for ' + key);
          break;
        case '':
          assert_array_equals([window.pageXOffset, window.pageYOffset], [0, 0], 'scroll is not restored before popstate for ' + key);
          t.done();
          return;
        default:
          assert_unreached();
          return;
      }

      window.history.back();
    }));

  }, 'Traversing history should restore scroll position before dispatching popstate');
</script>
